feat: move new apis to ee dir
