+++ lilypad/ee/server/schemas/__init__.py
@@ -1,5 +1,15 @@
 """The module for the `lilypad` EE schemas."""
 
 from .annotations import AnnotationCreate, AnnotationPublic, AnnotationUpdate
+from .deployments import DeploymentCreate, DeploymentPublic
+from .environments import EnvironmentCreate, EnvironmentPublic
 
-__all__ = ["AnnotationCreate", "AnnotationPublic", "AnnotationUpdate"]
+__all__ = [
+    "AnnotationCreate",
+    "AnnotationPublic",
+    "AnnotationUpdate",
+    "DeploymentCreate",
+    "DeploymentPublic",
+    "EnvironmentCreate",
+    "EnvironmentPublic",
+]
