+++ lilypad/ee/server/services/__init__.py
@@ -1,3 +1,5 @@
 """EE Services for the `lilypad` server."""
 
 from .annotations_service import AnnotationService
+from .deployments import DeploymentService
+from .environments import EnvironmentService
