+++ lilypad/ee/server/services/deployments.py
@@ -6,9 +6,10 @@ from uuid import UUID
 from fastapi import HTTPException, status
 from sqlmodel import desc, select
 
-from ..models import DeploymentTable, GenerationTable
-from ..schemas import DeploymentCreate
-from .base_organization import BaseOrganizationService
+from lilypad.ee.server.models.deployments import DeploymentTable
+from lilypad.ee.server.schemas.deployments import DeploymentCreate
+from lilypad.server.models import GenerationTable
+from lilypad.server.services.base_organization import BaseOrganizationService
 
 
 class DeploymentService(BaseOrganizationService[DeploymentTable, DeploymentCreate]):
