+++ lilypad/server/api/v0/main.py
@@ -9,7 +9,6 @@ from ....ee.server.api import v0_ee_api
 from ...settings import Settings, get_settings
 from .api_keys_api import api_keys_api
 from .auth import auth_router
-from .environment_api import environment_router
 from .generations_api import generations_router
 from .organizations_api import organization_router
 from .projects_api import projects_router
@@ -21,7 +20,6 @@ api = FastAPI(separate_input_output_schemas=False)
 # The `/ee` FastAPI sub-app for `lilypad`.
 api.mount("/ee", v0_ee_api)
 api.include_router(api_keys_api)
-api.include_router(environment_router)
 api.include_router(generations_router)
 api.include_router(projects_router)
 api.include_router(spans_router)
