+++ lilypad/server/models/__init__.py
@@ -3,8 +3,6 @@
 from .api_keys import APIKeyTable
 from .base_organization_sql_model import BaseOrganizationSQLModel
 from .base_sql_model import BaseSQLModel, JSONTypeDecorator, get_json_column
-from .deployments import DeploymentTable
-from .environments import EnvironmentTable
 from .generations import (
     GenerationTable,
     GenerationUpdate,
@@ -25,8 +23,6 @@ __all__ = [
     "APIKeyTable",
     "BaseOrganizationSQLModel",
     "BaseSQLModel",
-    "EnvironmentTable",
-    "DeploymentTable",
     "GenerationTable",
     "GenerationUpdate",
     "JSONTypeDecorator",
