+++ lilypad/server/services/__init__.py
@@ -1,8 +1,6 @@
 """Services for the `lilypad` server."""
 
 from .api_keys import APIKeyService
-from .deployments import DeploymentService
-from .environments import EnvironmentService
 from .generations import GenerationService
 from .organization_invites import OrganizationInviteService
 from .organizations import OrganizationService
